Remove textures/types
