+++ types/three/src/textures/DataTexture.d.ts
@@ -7,7 +7,6 @@ import {
     Wrapping,
 } from "../constants.js";
 import { Texture } from "./Texture.js";
-import { TextureImageData } from "./types.js";
 
 /**
  * Creates a texture directly from raw data, width and height.
@@ -109,4 +108,10 @@ export class DataTexture extends Texture {
      * @defaultValue `1`
      */
     unpackAlignment: number;
+}
+
+export interface TextureImageData {
+    data: Uint8Array | Uint8ClampedArray;
+    height: number;
+    width: number;
 }
