--- types/three/src/textures/types.d.ts
@@ -1,9 +1,0 @@
-export interface TextureImageData {
-    data: Uint8Array | Uint8ClampedArray;
-    height: number;
-    width: number;
-}
-
-export interface Texture3DImageData extends TextureImageData {
-    depth: number;
-}
